Further refactor using the strategy pattern in the Controller.
* This necessitates the use of instance_exec so we can break from the
loop by raising a StopIteration error (instead of the regular break).
* There's a lot going on here. We have our commands in a lookup table,
which is nice and extensible. We've totally separated the takeaway
implementation from this Controller layer. And we've pulled the command
variable out to a private method, clearly defining a command as just the
'first part of a given input string'.